Added automatic extraction on html apps
