+++ contentcuration/contentcuration/static/less/file-uploader.less
@@ -162,7 +162,7 @@
         .modal-header{ background-color:@gray-200; }
         #thumbnail_area{
             #generate_thumbnail_text, #generate_thumbnail_error{ display:none; font-weight: bold; }
-            #generate_thumbnail_error{ color:@red-error-color; margin-top: 15%; }
+            #generate_thumbnail_error{ color:@red-error-color; margin-top: 15%; word-wrap: break-word;}
             #thumbnail_buttons{ padding-bottom: 20px; }
             #thumbnail_preview{
                 height:@thumbnail-modal-height * 0.7;
