+++ contentcuration/contentcuration/urls.py
@@ -113,8 +113,6 @@ urlpatterns = [
     url(r'^api/publish_channel/$', views.publish_channel, name='publish_channel'),
     url(r'^api/generate_thumbnail/$', views.generate_thumbnail, name='generate_thumbnail'),
     url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    url(r'exercises/$', views.exercise_list, name='exercise_list'),
-    url(r'exercises/(?P<exercise_id>\w+)', views.exercise, name='exercise'),
     url(r'^file_upload/', views.file_upload, name="file_upload"),
     url(r'^file_create/', views.file_create, name="file_create"),
     url(r'^channels/$', views.channel_list, name='channels'),
